Success
Backend has successfully been passed into the frontend and the front end is recieving data. Problem of how to make those images display is still somewhat of a question but I will be researching

Backend is passed completely through and accessable by the frontend. Frontend work to be done consists of like... figuing out how to display the images. Making the pixelart for the stuff, and going from there.

The backend still needs a LOT of cleanup work, and also to be integrated with the board pieces created when the program is running. All in all huge progress. And we will end up with something more complete here soon hopefully